Updated to allow building under MSYS2/MinGW
To get the code to build successfully under MinGW the following changes
were made.

The configure command for some of the projects had to be changed from
absolute paths to relative paths to avoid build errors relating to path
name issues under windows.

GMP, MPC and MPFR are built as part of GCC.

fd2sfd has a patch for it to build under MinGW.

sfdc is patched to remove carrage returns from the sfd files. This was
causing errors in the generated include files.