Add patch for CAMD to fix incorrect func definition
The .fd file supplied with the CAMD SDK contains an incorrect function
definition for CreateMidiA(), which means that a stub is not generated
for the CreateMidi() function.

This also causes fd2sfd to emit the warning:
Warning: CreateMidiA gets 2 params and 1 regs!

Add a patch to fix the warning from fd2sfd and allow sfdc to generate a
stub for CreateMidi().